full refactor to bomb
