+++ ophys-mfish-dev/bomb/data_files/behavior_stimulus_file.py
