+++ ophys-mfish-dev/bomb/processing/stimulus/utils.py
