+++ ophys-mfish-dev/bomb/processing/sync/sync_utilities.py
@@ -3,7 +3,7 @@ from typing import Tuple, Optional, List
 import numpy as np
 import pandas as pd
 
-from data_objects.sync.sync_dataset import SyncDataset
+from bomb.processing.sync.sync_dataset import SyncDataset
 
 """
 Copied from AllenSDK (version = ) on 01/19/2023 by @mattjdavis
