+++ ophys-mfish-dev/bomb/processing/timestamps/stimulus_timestamps/timestamps_processing.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 import numpy as np
 
-from allensdk.internal.brain_observatory.time_sync import OphysTimeAligner
+# from allensdk.internal.brain_observatory.time_sync import OphysTimeAligner
 
 
 def get_behavior_stimulus_timestamps(stimulus_pkl: dict) -> np.ndarray:
@@ -28,25 +28,25 @@ def get_behavior_stimulus_timestamps(stimulus_pkl: dict) -> np.ndarray:
     return stimulus_timestamps
 
 
-def get_ophys_stimulus_timestamps(sync_path: Union[str, Path]) -> np.ndarray:
-    """Obtain visual behavior stimuli timing information from a sync *.h5 file.
+# def get_ophys_stimulus_timestamps(sync_path: Union[str, Path]) -> np.ndarray:
+#     """Obtain visual behavior stimuli timing information from a sync *.h5 file.
 
-    Parameters
-    ----------
-    sync_path : Union[str, Path]
-        The path to a sync *.h5 file that contains global timing information
-        about multiple data streams (e.g. behavior, ophys, eye_tracking)
-        during a session.
+#     Parameters
+#     ----------
+#     sync_path : Union[str, Path]
+#         The path to a sync *.h5 file that contains global timing information
+#         about multiple data streams (e.g. behavior, ophys, eye_tracking)
+#         during a session.
 
-    Returns
-    -------
-    np.ndarray
-        Timestamps (in seconds) for presented stimulus frames during a
-        behavior + ophys session.
-    """
-    aligner = OphysTimeAligner(sync_file=sync_path)
-    stimulus_timestamps, _ = aligner.clipped_stim_timestamps
-    return stimulus_timestamps
+#     Returns
+#     -------
+#     np.ndarray
+#         Timestamps (in seconds) for presented stimulus frames during a
+#         behavior + ophys session.
+#     """
+#     aligner = OphysTimeAligner(sync_file=sync_path)
+#     stimulus_timestamps, _ = aligner.clipped_stim_timestamps
+#     return stimulus_timestamps
 
 
 def get_frame_indices(
