--- ophys-mfish-dev/data_objects/data_object.py
@@ -1,177 +1,0 @@
-import abc
-from collections import deque
-from enum import Enum
-from typing import Any, Optional, Set
-
-from allensdk.brain_observatory.comparison_utils import compare_fields
-
-
-class DataObject(abc.ABC):
-    """An abstract class that prototypes properties that represent a
-    category of experimental data/metadata (e.g. running speed,
-    rewards, licks, etc.) and that prototypes methods to allow conversion of
-    the experimental data/metadata to and from various
-    data sources and sinks (e.g. LIMS, JSON, NWB).
-    """
-
-    def __init__(self, name: str, value: Any,
-                 exclude_from_equals: Optional[Set[str]] = None,
-                 is_value_self: bool = False):
-        """
-        :param name
-            Name
-        :param value
-            Value
-        :param exclude_from_equals
-            Optional set which will exclude these properties from comparison
-            checks to another DataObject
-        :param is_value_self
-            Indicates that the `value` is `self`. Useful for representing
-            complex objects with multiple properties rather than a simple
-            array or dataframe
-        """
-        if value is self:
-            raise ValueError('Passing value of self is not supported')
-        if is_value_self and value is not None:
-            raise ValueError('If passing is_value_self=True, then value '
-                             'should be None')
-        self._name = name
-        self._value = value
-        self._is_value_self = is_value_self
-
-        efe = exclude_from_equals if exclude_from_equals else set()
-        self._exclude_from_equals = efe
-
-    @property
-    def name(self) -> str:
-        return self._name
-
-    @property
-    def value(self) -> Any:
-        if self._is_value_self:
-            value = self
-        else:
-            value = self._value
-        return value
-
-    def to_dict(self) -> dict:
-        """
-        Serialize DataObject to dict
-        :return
-            A nested dict serializing the DataObject
-
-        notes
-            If a DataObject contains properties, these properties will either:
-                1) be serialized to nested dict with "name" attribute of
-                    DataObject if the property is itself a DataObject
-                2) Value for property will be added with name of property
-        :examples
-            >>> class Simple(DataObject):
-            ...     def __init__(self):
-            ...         super().__init__(name='simple', value=1)
-            >>> s = Simple()
-            >>> assert s.to_dict() == {'simple': 1}
-
-            >>> class B(DataObject):
-            ...     def __init__(self):
-            ...         super().__init__(name='b', value='!')
-
-            >>> class A(DataObject):
-            ...     def __init__(self, b: B):
-            ...         super().__init__(name='a', value=None,
-            is_value_self=True)
-            ...         self._b = b
-            ...     @property
-            ...     def prop1(self):
-            ...         return self._b
-            ...     @property
-            ...     def prop2(self):
-            ...         return '@'
-            >>> a = A(b=B())
-            >>> assert a.to_dict() == {'a': {'b': '!', 'prop2': '@'}}
-        """
-        res = dict()
-        q = deque([(self._name, self, [])])
-
-        while q:
-            name, value, path = q.popleft()
-            if isinstance(value, DataObject):
-                # The path stores the nested key structure
-                # Here, build onto the nested key structure
-                newpath = path + [name]
-
-                def _get_keys_and_values(base_value: DataObject):
-                    properties = []
-                    for name, value in base_value._get_properties().items():
-                        if value is base_value:
-                            # skip properties that return self
-                            # (leads to infinite recursion)
-                            continue
-                        if name == 'name':
-                            # The name is the key
-                            continue
-
-                        if isinstance(value, DataObject):
-                            # The key will be the DataObject "name" field
-                            name = value._name
-                        else:
-                            # The key will be the property name
-                            pass
-                        properties.append((name, value, newpath))
-                    return properties
-                properties = _get_keys_and_values(base_value=value)
-
-                # Find the nested dict
-                cur = res
-                for p in path:
-                    cur = cur[p]
-
-                if isinstance(value._value, DataObject) or \
-                        value._is_value_self:
-                    # it's nested
-                    cur[value._name] = dict()
-                    for p in properties:
-                        q.append(p)
-                else:
-                    # it's flat
-                    cur[name] = value._value
-
-            else:
-                cur = res
-                for p in path:
-                    cur = cur[p]
-
-                if isinstance(value, Enum):
-                    # convert to string
-                    value = value.value
-                cur[name] = value
-
-        return res
-
-    def _get_properties(self):
-        """Returns all property names and values"""
-        def is_prop(attr):
-            return isinstance(getattr(type(self), attr, None), property)
-        props = [attr for attr in dir(self) if is_prop(attr)]
-        return {name: getattr(self, name) for name in props}
-
-    def __eq__(self, other: "DataObject"):
-        if type(self) != type(other):
-            msg = f'Do not know how to compare with type {type(other)}'
-            raise NotImplementedError(msg)
-
-        d_self = self.to_dict()
-        d_other = other.to_dict()
-
-        for p in d_self:
-            if p in self._exclude_from_equals:
-                continue
-            x1 = d_self[p]
-            x2 = d_other[p]
-
-            try:
-                compare_fields(x1=x1, x2=x2,
-                               ignore_keys=self._exclude_from_equals)
-            except AssertionError:
-                return False
-        return True
