--- ophys-mfish-dev/data_objects/timestamps/ophys_timestamps.py
@@ -1,105 +1,0 @@
-import logging
-
-import numpy as np
-from pynwb import NWBFile
-
-from allensdk.brain_observatory.behavior.data_files import SyncFile
-from allensdk.core import DataObject
-from allensdk.core import \
-    NwbReadableInterface
-from allensdk.brain_observatory.behavior.data_files.sync_file import \
-    SyncFileReadableInterface
-
-
-class OphysTimestamps(DataObject, SyncFileReadableInterface,
-                      NwbReadableInterface):
-    _logger = logging.getLogger(__name__)
-
-    def __init__(self, timestamps: np.ndarray):
-        """
-        :param timestamps
-            ophys timestamps
-        """
-        super().__init__(name='ophys_timestamps', value=timestamps)
-
-    @classmethod
-    def from_sync_file(cls, sync_file: SyncFile) -> "OphysTimestamps":
-        ophys_timestamps = sync_file.data['ophys_frames']
-        return cls(timestamps=ophys_timestamps)
-
-    @classmethod
-    def from_nwb(cls, nwbfile: NWBFile) -> "OphysTimestamps":
-        ts = nwbfile.processing[
-                 'ophys'].get_data_interface('dff').roi_response_series[
-                 'traces'].timestamps[:]
-        return cls(timestamps=ts)
-
-    def validate(self, number_of_frames: int) -> "OphysTimestamps":
-        """Validates that number of ophys timestamps do not exceed number of
-        dff traces. If so, truncates number of ophys timestamps to the same
-        length as dff traces
-
-        :param number_of_frames
-            number of frames in the movie
-
-        Notes
-        ---------
-        Modifies self._value if ophys timestamps exceed length of
-        number_of_frames
-        """
-        # Scientifica data has extra frames in the sync file relative
-        # to the number of frames in the video. These sentinel frames
-        # should be removed.
-        # NOTE: This fix does not apply to mesoscope data.
-        # See http://confluence.corp.alleninstitute.org/x/9DVnAg
-        ophys_timestamps = self.value
-        num_of_timestamps = len(ophys_timestamps)
-        if number_of_frames < num_of_timestamps:
-            self._logger.info(
-                "Truncating acquisition frames ('ophys_frames') "
-                f"(len={num_of_timestamps}) to the number of frames "
-                f"in the df/f trace ({number_of_frames}).")
-            self._value = ophys_timestamps[:number_of_frames]
-        elif number_of_frames > num_of_timestamps:
-            raise RuntimeError(
-                f"dff_frames (len={number_of_frames}) is longer "
-                f"than timestamps (len={num_of_timestamps}).")
-        return self
-
-
-class OphysTimestampsMultiplane(OphysTimestamps):
-    def __init__(self, timestamps: np.ndarray):
-        super().__init__(timestamps=timestamps)
-
-    @classmethod
-    def from_sync_file(cls, sync_file: SyncFile,
-                       group_count: int,
-                       plane_group: int) -> "OphysTimestampsMultiplane":
-        if group_count == 0:
-            raise ValueError('Group count cannot be 0')
-
-        ophys_timestamps = sync_file.data['ophys_frames']
-        cls._logger.info(
-            "Mesoscope data detected. Splitting timestamps "
-            f"(len={len(ophys_timestamps)} over {group_count} "
-            "plane group(s).")
-
-        # Resample if collecting multiple concurrent planes
-        # because the frames are interleaved
-        ophys_timestamps = ophys_timestamps[plane_group::group_count]
-
-        return cls(timestamps=ophys_timestamps)
-
-    def validate(self, number_of_frames: int) -> "OphysTimestampsMultiplane":
-        """
-        Raises error if length of timestamps and number of frames are not equal
-        :param number_of_frames
-            See super().validate
-        """
-        ophys_timestamps = self.value
-        num_of_timestamps = len(ophys_timestamps)
-        if number_of_frames != num_of_timestamps:
-            raise RuntimeError(
-                f"dff_frames (len={number_of_frames}) is not equal to "
-                f"number of split timestamps (len={num_of_timestamps}).")
-        return self
